修复异步调度中的部分问题 (merge request !1379)
TAPD: --story=127699001

+++ src/ksana_llm/runtime/forward_request.h
@@ -25,7 +25,7 @@ enum class ForwardRequestType : uint8_t {
 // The information used for forward.
 struct ForwardRequest {
   // The request id.
-  size_t req_id;
+  int64_t req_id;
 
   // The infer stage, context decode or decode.
   InferStage infer_stage = InferStage::STAGE_CONTEXT;
